+++ Client/Animation.h
@@ -52,7 +52,7 @@ public:
 	const wstring& GetAnimName() { return _animationName; }
 	Animator* GetAnimator() { return p_animator; }
 	AnimFrame& GetAnimFrame(int idx) { return _vecAnimFrame[idx]; }
-	int GetMaxFrame() { return _vecAnimFrame.size(); }
+	UINT GetMaxFrame() { return (UINT)_vecAnimFrame.size(); }
 
 	friend class Animator;
 };
